+++ android/src/main/kotlin/be/appmire/flutter_keychain/StringEncryptor.kt
@@ -1,0 +1,9 @@
+package be.appmire.flutter_keychain
+
+interface StringEncryptor {
+    @Throws(Exception::class)
+    fun encrypt(input: String?): String?
+
+    @Throws(Exception::class)
+    fun decrypt(input: String?): String?
+}