+++ example/ios/RunnerTests/RunnerTests.swift
@@ -1,0 +1,31 @@
+import Flutter
+import UIKit
+import XCTest
+
+@testable import flutter_keychain
+
+// This demonstrates a simple unit test of the Swift portion of this plugin's implementation.
+//
+// See https://developer.apple.com/documentation/xctest for more information about using XCTest.
+
+class RunnerTests: XCTestCase {
+
+  // TODO: fix tests for get/put/remove/clear
+
+  /*
+  func testGetPlatformVersion() {
+    let plugin = FlutterKeychainPlugin()
+
+    let call = FlutterMethodCall(methodName: "getPlatformVersion", arguments: [])
+
+    let resultExpectation = expectation(description: "result block must be called.")
+    plugin.handle(call) { result in
+      XCTAssertEqual(result as! String, "iOS " + UIDevice.current.systemVersion)
+      resultExpectation.fulfill()
+    }
+    waitForExpectations(timeout: 1)
+  }
+
+  */
+
+}