+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.3.1
+
+* Raise minimum Flutter version to 3.3.0
+* Add support for specifying the keychain name on Android & iOS.
+
 ## 2.3.0
 
 * Update dependencies, gradle